hello_world/requirements.txt
requests
flask-lambda-python36 